Abstract code function to new class
