+++ src/main/java/hirehive/address/logic/commands/FindCommand.java
@@ -3,7 +3,6 @@ package hirehive.address.logic.commands;
 import static java.util.Objects.requireNonNull;
 
 import hirehive.address.commons.util.ToStringBuilder;
-import hirehive.address.logic.Messages;
 import hirehive.address.model.Model;
 import hirehive.address.model.person.NameContainsKeywordsPredicate;
 
@@ -28,7 +27,7 @@ public class FindCommand extends Command {
 
     @Override
     public CommandResult execute(Model model) {
-        return FilterUtil.executeFilter(model, predicate);
+        return CommandUtil.executeFilter(model, predicate);
         /*
         requireNonNull(model);
         model.updateFilteredPersonList(predicate);
